x86/gsseg: Move load_gs_index() to its own new header file
commit ae53fa18703000f507107df43efd1168a0365361 upstream.

GS is a special segment on x86_64, move load_gs_index() to its own new
header file to simplify header inclusion.

No change in functionality.

Intel-SIG: commit ae53fa187030 x86/gsseg: Move load_gs_index() to its
own new header file.

Link: https://lore.kernel.org/r/20230112072032.35626-5-xin3.li@intel.com

Conflicts:
    arch/x86/include/asm/gsseg.h
    arch/x86/include/asm/special_insns.h
[ Xingjiang Lu: resolve context conflicts ]
